+++ manager/app/src/main/cpp/ksu.h
@@ -12,9 +12,9 @@
 
 #include "uapi/ksu.h"
 
-uint16_t get_kernel_uapi_version();
+uint32_t get_kernel_uapi_version();
 
-uint16_t get_manager_uapi_version();
+uint32_t get_manager_uapi_version();
 
 uint32_t get_version();
